Don't use the cname cache when using DnsRecordResolveContext
Motivation:

When we resolve other records then AAAA / AAAA we will use the resolve(...) overloads that take a DnsQuestion. This will then end up creating a DnsRecordResolveContext, in this case we should not use the CnameCache at all as we also not the result.

Modifications:

Don't cache CNAME and also dont use the cache for CNAME when using DnsRecordResolveContext

Result:

More correct resolving and also not possible to have failures due CNAME still be in the cache while the queried record experied